@@ -18,13 +18,8 @@ import { Helpers as projectHelpers } from '../project/helpers';
 import { getEnvVarsByProjectId } from '../env-variables/resolvers';
 import { logger } from '../../loggers/logger';
 
-export const getRestoreLocation: ResolverFn = async (
-  restore,
-  args,
-  context,
-) => {
-  const { restoreLocation, backupId } = restore;
-  const { sqlClientPool, hasPermission } = context;
+const getRestoreLocation = async (backupId, restoreLocation, sqlClientPool) => {
+  let restoreSize = 0;
   const rows = await query(sqlClientPool, Sql.selectBackupByBackupId(backupId));
   const project = await projectHelpers(sqlClientPool).getProjectByEnvironmentId(rows[0].environment);
   const projectEnvVars = await query(sqlClientPool, Sql.selectEnvVariablesByProjectsById(project.projectId));
@@ -100,44 +95,36 @@ export const getRestoreLocation: ResolverFn = async (
 
 
     // before generating the signed url, check the object exists
-    let exists = false
     const restoreLoc = await s3Client.headObject({
       Bucket: R.prop(2, s3Parts),
       Key: R.prop(3, s3Parts)
     });
     try {
-      await Promise.all([restoreLoc.promise()]).then(data => {
-        // the file exists
-      }).catch(err => {
-        if (err) throw err;
-      });
-      exists = true
-    } catch(err) {
-      exists = false
-    }
-    if (exists) {
-      return s3Client.getSignedUrl('getObject', {
+      const data = await Promise.resolve(restoreLoc.promise());
+      restoreSize = data.ContentLength
+      const restLoc = await s3Client.getSignedUrl('getObject', {
         Bucket: R.prop(2, s3Parts),
         Key: R.prop(3, s3Parts),
         Expires: 300 // 5 minutes
-      });
-    } else {
+      })
+      return [restLoc, restoreSize];
+    } catch(err) {
       await query(
         sqlClientPool,
         Sql.deleteRestore({
           backupId
         })
       );
-      return ""
+      return ["", restoreSize];
     }
   }
 
-  return restoreLocation;
+  return [restoreLocation, restoreSize];
 };
 
 export const getBackupsByEnvironmentId: ResolverFn = async (
   { id: environmentId },
-  { includeDeleted, limit },
+  { limit },
   { sqlClientPool, hasPermission, adminScopes }
 ) => {
   const environment = await environmentHelpers(
@@ -154,10 +141,6 @@ export const getBackupsByEnvironmentId: ResolverFn = async (
     .where('environment', environmentId)
     .orderBy('created', 'desc')
     .orderBy('id', 'desc');
-
-  if (!includeDeleted) {
-    queryBuilder = queryBuilder.where('deleted', '0000-00-00 00:00:00');
-  }
 
   if (limit) {
     queryBuilder = queryBuilder.limit(limit);
@@ -178,16 +161,28 @@ export const addBackup: ResolverFn = async (
     project: environment.project
   });
 
-  const { insertId } = await query(
-    sqlClientPool,
-    Sql.insertBackup({
-      id,
-      environment: environmentId,
-      source,
-      backupId,
-      created
-    })
-  );
+  let insertId: number;
+  try {
+     ({insertId} = await query(
+      sqlClientPool,
+      Sql.insertBackup({
+        id,
+        environment: environmentId,
+        source,
+        backupId,
+        created
+      })
+    ));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error adding backup. Backup already exists.`
+      );
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(sqlClientPool, Sql.selectBackup(insertId));
   const backup = R.prop(0, rows);
 
@@ -259,16 +254,28 @@ export const addRestore: ResolverFn = async (
     project: R.path(['0', 'pid'], perms)
   });
 
-  const { insertId } = await query(
-    sqlClientPool,
-    Sql.insertRestore({
-      id,
-      backupId,
-      status,
-      restoreLocation,
-      created
-    })
-  );
+  let insertId: number;
+  try {
+     ({insertId} = await query(
+      sqlClientPool,
+      Sql.insertRestore({
+        id,
+        backupId,
+        status,
+        restoreLocation,
+        created
+      })
+    ));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error adding restore. Restore already exists.`
+      );
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   let rows = await query(sqlClientPool, Sql.selectRestore(insertId));
   const restoreData = R.prop(0, rows);
 
@@ -410,8 +417,13 @@ export const getRestoreByBackupId: ResolverFn = async (
     sqlClientPool,
     Sql.selectRestoreByBackupId(backupId)
   );
-
-  return R.prop(0, rows);
+  const row = R.prop(0, rows)
+  if (row && row.restoreLocation != null) {
+    // if the restore has a location, determine the signed url and the reported size of the object in Bytes
+    const [restLoc, restSize] = await getRestoreLocation(backupId, row.restoreLocation, sqlClientPool);
+    return {...row, restoreLocation: restLoc, restoreSize: restSize};
+  }
+  return row;
 };
 
 export const backupSubscriber = createEnvironmentFilteredSubscriber([
