@@ -21,11 +21,24 @@ const DISABLE_NON_ORGANIZATION_NOTIFICATION_ASSIGNMENT = process.env.DISABLE_NON
 
 const addNotificationGeneric = async (sqlClientPool, notificationTable, input) => {
   const createSql = knex(notificationTable).insert(input).toString();
-  let { insertId } = await query(sqlClientPool, createSql);
+
+  let insertId: number;
+  try {
+     ({insertId} = await query(sqlClientPool, createSql));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error adding notification ${input.name}. Notification already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   return await query(sqlClientPool, knex(notificationTable).where('id', insertId).toString());
 }
 
-const checkOrgNotificationPermission = async (hasPermission, input) => {
+const checkOrgNotificationPermission = async (hasPermission, input, adminScopes) => {
   if (input.organization != null) {
     const organizationData = await organizationHelpers(sqlClientPool).getOrganizationById(input.organization);
     if (organizationData === undefined) {
@@ -43,7 +56,7 @@ const checkOrgNotificationPermission = async (hasPermission, input) => {
       );
     }
   } else {
-    if (DISABLE_NON_ORGANIZATION_NOTIFICATION_ASSIGNMENT == "false") {
+    if (DISABLE_NON_ORGANIZATION_NOTIFICATION_ASSIGNMENT == "false" || adminScopes.projectViewAll) {
       await hasPermission('notification', 'add');
     } else {
       throw new Error(
@@ -56,41 +69,41 @@ const checkOrgNotificationPermission = async (hasPermission, input) => {
 export const addNotificationMicrosoftTeams: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission }
+  { sqlClientPool, hasPermission, adminScopes}
 ) => {
-  await checkOrgNotificationPermission(hasPermission, input)
+  await checkOrgNotificationPermission(hasPermission, input, adminScopes)
   return R.path([0], await addNotificationGeneric(sqlClientPool, 'notification_microsoftteams', input));
 };
 
 export const addNotificationEmail: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission }
+  { sqlClientPool, hasPermission, adminScopes}
 ) => {
-  await checkOrgNotificationPermission(hasPermission, input)
+  await checkOrgNotificationPermission(hasPermission, input, adminScopes)
   return R.path([0], await addNotificationGeneric(sqlClientPool, 'notification_email', input));
 };
 
 export const addNotificationRocketChat: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission }
+  { sqlClientPool, hasPermission, adminScopes }
 ) => {
-  await checkOrgNotificationPermission(hasPermission, input)
+  await checkOrgNotificationPermission(hasPermission, input, adminScopes)
   return R.path([0], await addNotificationGeneric(sqlClientPool, 'notification_rocketchat', input));
 };
 
 export const addNotificationSlack: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission }
+  { sqlClientPool, hasPermission, adminScopes}
 ) => {
-  await checkOrgNotificationPermission(hasPermission, input)
+  await checkOrgNotificationPermission(hasPermission, input, adminScopes)
   return R.path([0], await addNotificationGeneric(sqlClientPool, 'notification_slack', input));
 };
 
-export const addNotificationWebhook: ResolverFn = async (root, { input }, { sqlClientPool, hasPermission }) => {
-  await checkOrgNotificationPermission(hasPermission, input)
+export const addNotificationWebhook: ResolverFn = async (root, { input }, { sqlClientPool, hasPermission, adminScopes}) => {
+  await checkOrgNotificationPermission(hasPermission, input, adminScopes)
   return R.path([0], await addNotificationGeneric(sqlClientPool, 'notification_webhook', input));
 };
 
@@ -484,6 +497,12 @@ const checkNotificationUpdatePermissions = async (check, hasPermission) => {
   }
 }
 
+const checkNotificationExists = (name, check) => {
+  if (R.find(R.propEq('name', `${name}`))(check) == undefined) {
+    throw new Error(`No notification found for ${name}`);
+  }
+}
+
 export const updateNotificationMicrosoftTeams: ResolverFn = async (
   root,
   { input },
@@ -497,9 +516,23 @@ export const updateNotificationMicrosoftTeams: ResolverFn = async (
     sqlClientPool,
     Sql.selectNotificationMicrosoftTeamsByName(name)
   );
+
+  checkNotificationExists(name, check);
+
   await checkNotificationUpdatePermissions(check, hasPermission)
 
-  await query(sqlClientPool, Sql.updateNotificationMicrosoftTeams(input));
+  try {
+    await query(sqlClientPool, Sql.updateNotificationMicrosoftTeams(input));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error renaming notification ${input.name}. Notification ${input.patch.name} already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(
     sqlClientPool,
     Sql.selectNotificationMicrosoftTeamsByName(name)
@@ -524,9 +557,23 @@ export const updateNotificationWebhook: ResolverFn = async (
     sqlClientPool,
     Sql.selectNotificationWebhookByName(name)
   );
+
+  checkNotificationExists(name, check);
+
   await checkNotificationUpdatePermissions(check, hasPermission)
 
-  await query(sqlClientPool, Sql.updateNotificationWebhook(input));
+  try {
+    await query(sqlClientPool, Sql.updateNotificationWebhook(input));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error renaming notification ${input.name}. Notification ${input.patch.name} already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(
     sqlClientPool,
     Sql.selectNotificationWebhookByName(name),
@@ -548,9 +595,23 @@ export const updateNotificationEmail: ResolverFn = async (
     sqlClientPool,
     Sql.selectNotificationEmailByName(name)
   );
+
+  checkNotificationExists(name, check)
+
   await checkNotificationUpdatePermissions(check, hasPermission)
 
-  await query(sqlClientPool, Sql.updateNotificationEmail(input));
+  try {
+    await query(sqlClientPool, Sql.updateNotificationEmail(input));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error renaming notification ${input.name}. Notification ${input.patch.name} already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(
     sqlClientPool,
     Sql.selectNotificationEmailByName(name)
@@ -572,9 +633,23 @@ export const updateNotificationRocketChat: ResolverFn = async (
     sqlClientPool,
     Sql.selectNotificationRocketChatByName(name)
   );
+
+  checkNotificationExists(name, check)
+
   await checkNotificationUpdatePermissions(check, hasPermission)
 
-  await query(sqlClientPool, Sql.updateNotificationRocketChat(input));
+  try {
+    await query(sqlClientPool, Sql.updateNotificationRocketChat(input));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error renaming notification ${input.name}. Notification ${input.patch.name} already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(
     sqlClientPool,
     Sql.selectNotificationRocketChatByName(name)
@@ -596,9 +671,23 @@ export const updateNotificationSlack: ResolverFn = async (
     sqlClientPool,
     Sql.selectNotificationSlackByName(name)
   );
+
+  checkNotificationExists(name, check)
+
   await checkNotificationUpdatePermissions(check, hasPermission)
 
-  await query(sqlClientPool, Sql.updateNotificationSlack(input));
+  try {
+    await query(sqlClientPool, Sql.updateNotificationSlack(input));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error renaming notification ${input.name}. Notification ${input.patch.name} already exists`
+      )
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(
     sqlClientPool,
     Sql.selectNotificationSlackByName(name)
@@ -683,4 +772,16 @@ export const removeAllNotificationsFromAllProjects: ResolverFn = async (
 
   // TODO: Check rows for success
   return 'success';
+};
+
+export const getAllNotifications: ResolverFn = async (
+  root,
+  args,
+  { sqlClientPool, hasPermission }
+) => {
+  await hasPermission('notification', 'viewAll');
+
+  const rows = await Helpers(sqlClientPool).selectAllNotifications();
+
+  return rows;
 };
