@@ -1,4 +1,3 @@
-// @ts-ignore
 import * as R from 'ramda';
 import { ResolverFn } from '../';
 import { logger } from '../../loggers/logger';
@@ -9,8 +8,8 @@ import { Sql } from './sql';
 import { arrayDiff } from '../../util/func';
 import { Helpers as openshiftHelpers } from '../openshift/helpers';
 import { Helpers as notificationHelpers } from '../notification/helpers';
+import { Helpers as groupHelpers } from '../group/helpers';
 import validator from 'validator';
-import { log } from 'winston';
 
 const isValidName = value => {
   if (validator.matches(value, /[^0-9a-z-]/)) {
@@ -33,7 +32,10 @@ export const addOrganization: ResolverFn = async (
   isValidName(input.name)
 
   try {
-      await hasPermission('organization', 'add');
+    await hasPermission('organization', 'add');
+    const org = await query(sqlClientPool, Sql.selectOrganizationByName(input.name));
+    // if no organization found, create it
+    if (R.length(org) == 0) {
       const { insertId } = await query(sqlClientPool, Sql.insertOrganization(input));
       const rows = await query(sqlClientPool, Sql.selectOrganization(insertId));
 
@@ -49,8 +51,11 @@ export const addOrganization: ResolverFn = async (
       });
 
       return R.prop(0, rows);
+    } else {
+      throw new Error(`There was an error creating the organization, ${input.name} already exists`);
+    }
   }  catch (err) {
-      throw new Error(`There was an error creating the organization ${input.name} ${err}`);
+    throw new Error(`There was an error creating the organization ${input.name} ${err}`);
   }
 };
 
@@ -334,7 +339,7 @@ export const getNotificationsForOrganizationProjectId: ResolverFn = async (
 export const getOwnersByOrganizationId: ResolverFn = async (
   { id: oid },
   _input,
-  { hasPermission, models, keycloakGrant, keycloakUsersGroups }
+  { hasPermission, models }
 ) => {
   await hasPermission('organization', 'view', {
     organization: oid,
@@ -347,13 +352,13 @@ export const getOwnersByOrganizationId: ResolverFn = async (
 export const getGroupsByOrganizationId: ResolverFn = async (
   { id: oid },
   _input,
-  { hasPermission, models, keycloakGrant, keycloakGroups }
+  { hasPermission, models, sqlClientPool }
 ) => {
   await hasPermission('organization', 'viewGroup', {
     organization: oid,
   });
 
-  const orgGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(oid, keycloakGroups);
+  const orgGroups = await groupHelpers(sqlClientPool).selectGroupsByOrganizationId(models, oid)
 
   return orgGroups;
 };
@@ -362,13 +367,13 @@ export const getGroupsByOrganizationId: ResolverFn = async (
 export const getUsersByOrganizationId: ResolverFn = async (
   _,
   args,
-  { hasPermission, models, keycloakGrant, keycloakGroups }
+  { hasPermission, models, sqlClientPool }
 ) => {
   await hasPermission('organization', 'viewUsers', {
     organization: args.organization,
   });
 
-  const orgGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(args.organization, keycloakGroups);
+  const orgGroups = await groupHelpers(sqlClientPool).selectGroupsByOrganizationId(models, args.organization)
 
   let members = []
   for (const group in orgGroups) {
@@ -384,6 +389,7 @@ export const getUsersByOrganizationId: ResolverFn = async (
       if (!exists) {
         // quick check to set the owner flag or not
         groupMembers[member].user.owner = false
+        groupMembers[member].user.admin = false
         groupMembers[member].user.comment = null
         if (groupMembers[member].user.attributes["comment"]) {
           groupMembers[member].user.comment = groupMembers[member].user.attributes["comment"][0]
@@ -392,6 +398,13 @@ export const getUsersByOrganizationId: ResolverFn = async (
           for (const a in groupMembers[member].user.attributes["lagoon-organizations"]) {
             if (parseInt(groupMembers[member].user.attributes["lagoon-organizations"][a]) == args.organization) {
               groupMembers[member].user.owner = true
+            }
+          }
+        }
+        if (groupMembers[member].user.attributes["lagoon-organizations-admin"]) {
+          for (const a in groupMembers[member].user.attributes["lagoon-organizations-admin"]) {
+            if (parseInt(groupMembers[member].user.attributes["lagoon-organizations-admin"][a]) == args.organization) {
+              groupMembers[member].user.admin = true
             }
           }
         }
@@ -418,6 +431,7 @@ export const getUserByEmailAndOrganizationId: ResolverFn = async (
     const queryUserGroups = await models.UserModel.getAllGroupsForUser(user.id, organization);
 
     user.owner = false
+    user.admin = false
     user.comment = null
     if (user.attributes["comment"]) {
       user.comment = user.attributes["comment"][0]
@@ -426,6 +440,13 @@ export const getUserByEmailAndOrganizationId: ResolverFn = async (
       for (const a in user.attributes["lagoon-organizations"]) {
         if (parseInt(user.attributes["lagoon-organizations"][a]) == organization) {
           user.owner = true
+        }
+      }
+    }
+    if (user.attributes["lagoon-organizations-admin"]) {
+      for (const a in user.attributes["lagoon-organizations-admin"]) {
+        if (parseInt(user.attributes["lagoon-organizations-admin"][a]) == organization) {
+          user.admin = true
         }
       }
     }
@@ -444,7 +465,7 @@ export const getUserByEmailAndOrganizationId: ResolverFn = async (
 export const getGroupRolesByUserIdAndOrganization: ResolverFn =async (
   { id: uid, organization },
   _input,
-  { hasPermission, models, keycloakGrant, keycloakUsersGroups, adminScopes }
+  { hasPermission, models, adminScopes }
 ) => {
   if (organization) {
     const queryUserGroups = await models.UserModel.getAllGroupsForUser(uid, organization);
@@ -491,9 +512,9 @@ export const getGroupsByNameAndOrganizationId: ResolverFn = async (
 export const getGroupCountByOrganizationProject: ResolverFn = async (
   { id: pid },
   _input,
-  { sqlClientPool, models, keycloakGroups }
+  { sqlClientPool, models }
 ) => {
-  const orgProjectGroups = await models.GroupModel.loadGroupsByProjectIdFromGroups(pid, keycloakGroups);
+  const orgProjectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
   return orgProjectGroups.length
 }
 
@@ -504,50 +525,13 @@ export const getGroupCountByOrganizationProject: ResolverFn = async (
 export const getGroupsByOrganizationsProject: ResolverFn = async (
   { id: pid },
   _input,
-  { sqlClientPool, models, keycloakGrant, keycloakGroups, keycloakUsersGroups, adminScopes }
+  { sqlClientPool, models }
 ) => {
-  const orgProjectGroups = await models.GroupModel.loadGroupsByProjectIdFromGroups(pid, keycloakGroups);
-  if (adminScopes.projectViewAll) {
-    // if platform owner, this will show ALL groups on a project (those that aren't in the organization too, yes its possible with outside intervention :| )
-    return orgProjectGroups;
-  }
-
-  const user = await models.UserModel.loadUserById(
-    keycloakGrant.access_token.content.sub
-  );
-  // if this user is an owner of an organization, then also display org based groups to this user
-  // when listing project groups
-  const userGroups = keycloakUsersGroups;
-  const usersOrgs = R.defaultTo('', R.prop('lagoon-organizations',  user.attributes)).toString()
-  const usersOrgsViewer = R.defaultTo('', R.prop('lagoon-organizations-viewer',  user.attributes)).toString()
-
-  if (usersOrgs != "" ) {
-    const usersOrgsArr = usersOrgs.split(',');
-    for (const userOrg of usersOrgsArr) {
-      const project = await projectHelpers(sqlClientPool).getProjectById(pid);
-      if (project.organization == userOrg) {
-        const orgGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(project.organization, keycloakGroups);
-        for (const pGroup of orgGroups) {
-          userGroups.push(pGroup)
-        }
-      }
-    }
-  }
-  if (usersOrgsViewer != "" ) {
-    const usersOrgsArr = usersOrgsViewer.split(',');
-    for (const userOrg of usersOrgsArr) {
-      const project = await projectHelpers(sqlClientPool).getProjectById(pid);
-      if (project.organization == userOrg) {
-        const orgViewerGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(project.organization, keycloakGroups);
-        for (const pGroup of orgViewerGroups) {
-          userGroups.push(pGroup)
-        }
-      }
-    }
-  }
-  const userProjectGroups = R.intersection(orgProjectGroups, userGroups);
-
-  return userProjectGroups;
+  // rather than make this more complicated than it needs to be, just return all the groups attached to a project, even those that aren't in the organization
+  // this way the organization owner can make the decision to remove the group from the project
+  // this isn't an impossibile situation, but is unlikely, due to the historical nature of lagoons group attachment process
+  const orgProjectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
+  return orgProjectGroups;
 };
 
 // check an existing project and the associated groups can be added to an organization
@@ -611,7 +595,7 @@ const checkProjectGroupAssociation = async (oid, projectGroups, projectGroupName
 export const getProjectGroupOrganizationAssociation: ResolverFn = async (
   _root,
   { input },
-  { sqlClientPool, models, hasPermission, keycloakGroups }
+  { sqlClientPool, models, hasPermission }
 ) => {
   let pid = input.project;
   let oid = input.organization;
@@ -625,7 +609,7 @@ export const getProjectGroupOrganizationAssociation: ResolverFn = async (
   const otherOrgs = []
 
   // get all the groups the requested project is in
-  const projectGroups = await models.GroupModel.loadGroupsByProjectIdFromGroups(pid, keycloakGroups);
+  const projectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
   await checkProjectGroupAssociation(oid, projectGroups, projectGroupNames, otherOrgs, groupProjectIds, projectInOtherOrgs, sqlClientPool)
 
   return "success";
@@ -637,7 +621,7 @@ export const getProjectGroupOrganizationAssociation: ResolverFn = async (
 export const removeProjectFromOrganization: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission, models, keycloakGroups, userActivityLogger }
+  { sqlClientPool, hasPermission, models, userActivityLogger }
 ) => {
   // platform admin only
   await hasPermission('organization', 'add');
@@ -651,7 +635,7 @@ export const removeProjectFromOrganization: ResolverFn = async (
   }
 
   try {
-    const projectGroups = await models.GroupModel.loadGroupsByProjectIdFromGroups(pid, keycloakGroups);
+    const projectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
 
     let removeGroups = []
     for (const g in projectGroups) {
@@ -720,27 +704,60 @@ export const removeProjectFromOrganization: ResolverFn = async (
 }
 
 // add existing project to an organization
+// this removes all notifications and groups from the project and resets all the access to the project to only
+// the default project group and the default-user of the project
 export const addExistingProjectToOrganization: ResolverFn = async (
   root,
   { input },
-  { sqlClientPool, hasPermission, userActivityLogger, models, keycloakGroups }
+  { sqlClientPool, hasPermission, userActivityLogger, models }
 ) => {
-
-  let pid = input.project;
-  let oid = input.organization;
 
   // platform admin only as it potentially reveals information about projects/orgs/groups
   await hasPermission('organization', 'add');
 
-  const groupProjectIds = []
-  const projectInOtherOrgs = []
-  const projectGroupNames = []
-  const otherOrgs = []
+  let pid = input.project;
+  let oid = input.organization;
+  const project = await projectHelpers(sqlClientPool).getProjectById(pid)
 
   // get all the groups the requested project is in
-  const projectGroups = await models.GroupModel.loadGroupsByProjectIdFromGroups(pid, keycloakGroups);
-  await checkProjectGroupAssociation(oid, projectGroups, projectGroupNames, otherOrgs, groupProjectIds, projectInOtherOrgs, sqlClientPool)
+  try {
+    const projectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
 
+    let removeGroups = []
+    for (const g in projectGroups) {
+      if (projectGroups[g].attributes["type"] == "project-default-group") {
+        // remove all users from the project default group except the `default-user@project`
+        await models.GroupModel.removeNonProjectDefaultUsersFromGroup(projectGroups[g], project.name)
+        // update group
+        await models.GroupModel.updateGroup({
+          id: projectGroups[g].id,
+          name: projectGroups[g].name,
+          attributes: {
+            ...projectGroups[g].attributes,
+            "lagoon-organization": [""]
+          }
+        });
+      } else {
+        removeGroups.push(projectGroups[g])
+      }
+    }
+    // remove groups from project
+    await models.GroupModel.removeProjectFromGroups(pid, removeGroups);
+  } catch (err) {
+    throw new Error(
+      `Unable to remove all groups from the project`
+    )
+  }
+  try {
+    // remove all notifications from project
+    await notificationHelpers(sqlClientPool).removeAllNotificationsFromProject({project: pid})
+  } catch (err) {
+    throw new Error(
+      `Unable to remove all notifications from the project`
+    )
+  }
+
+  const projectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
   // check if project.organization is already set?
   // get all groups the project is in
   for (const group of projectGroups) {
@@ -753,6 +770,7 @@ export const addExistingProjectToOrganization: ResolverFn = async (
         "lagoon-organization": [input.organization]
       }
     });
+    await groupHelpers(sqlClientPool).addOrganizationToGroup(input.organization, group.id)
 
     // log this activity
     userActivityLogger(`User added a group to organization`, {
@@ -781,7 +799,7 @@ export const addExistingProjectToOrganization: ResolverFn = async (
   );
 
   // log this activity
-  userActivityLogger(`User added a project to organization`, {
+  userActivityLogger(`User added an existing project to organization`, {
     project: '',
     organization: input.organization,
     event: 'api:addExistingProjectToOrganization',
@@ -911,7 +929,7 @@ export const addExistingGroupToOrganization: ResolverFn = async (
 export const removeUserFromOrganizationGroups: ResolverFn = async (
   _root,
   { input: { user: userInput, organization: organizationInput } },
-  { models, sqlClientPool, hasPermission, keycloakGroups, userActivityLogger }
+  { models, sqlClientPool, hasPermission, userActivityLogger }
 ) => {
 
   if (R.isEmpty(userInput)) {
@@ -920,7 +938,7 @@ export const removeUserFromOrganizationGroups: ResolverFn = async (
 
   const user = await models.UserModel.loadUserByIdOrUsername({
     id: R.prop('id', userInput),
-    username: R.prop('email', userInput)
+    email: R.prop('email', userInput)
   });
 
   // check the organization exists
@@ -933,7 +951,7 @@ export const removeUserFromOrganizationGroups: ResolverFn = async (
   await hasPermission('organization', 'removeGroup', {
     organization: organizationInput,
   });
-  const orgGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(organizationInput, keycloakGroups);
+  const orgGroups = await groupHelpers(sqlClientPool).selectGroupsByOrganizationId(models, organizationInput)
 
   // iterate through groups and remove the user
   let groupsRemoved = []
@@ -969,7 +987,7 @@ export const removeUserFromOrganizationGroups: ResolverFn = async (
 export const deleteOrganization: ResolverFn = async (
   _root,
   { input },
-  { sqlClientPool, hasPermission, userActivityLogger, models, keycloakGroups }
+  { sqlClientPool, hasPermission, userActivityLogger, models }
 ) => {
   await hasPermission('organization', 'delete', {
     organization: input.id
@@ -1002,7 +1020,7 @@ export const deleteOrganization: ResolverFn = async (
     );
   }
 
-  const orgGroups = await models.GroupModel.loadGroupsByOrganizationIdFromGroups(orgResult.id, keycloakGroups);
+  const orgGroups = await groupHelpers(sqlClientPool).selectGroupsByOrganizationId(models, orgResult.id)
   if (orgGroups.length > 0) {
     // throw error if there are any existing environments
     throw new Error(
@@ -1037,3 +1055,247 @@ export const deleteOrganization: ResolverFn = async (
   });
   return 'success';
 };
+
+// recursive check of a project and the groups of that project
+const checkProjectGroups = async (groupProjectIds, projectIds, projectsGroups, models, sqlClientPool, pid) => {
+  const projectGroups = await groupHelpers(sqlClientPool).selectGroupsByProjectId(models, pid)
+  let index = projectIds.findIndex((item) => item === pid);
+  if (index === -1) {
+    projectIds.push(pid)
+    // get all the groups the requested project is in
+    for (const group of projectGroups) {
+      const groupProjectIdss = await groupHelpers(sqlClientPool).selectProjectIdsByGroupID(group.id)
+      // for each group the project is in, get the list of projects that are also in this group
+      for (const project of groupProjectIdss) {
+        let index = groupProjectIds.findIndex((item) => item.group === group.name);
+        if (index === -1) {
+          groupProjectIds.push({group: group.name, project: project})
+          projectsGroups.push(group)
+        }
+        // recurse the project if it hasn't already had been done
+        let index2 = projectIds.findIndex((item) => item === project);
+        if (index2 === -1) {
+          await checkProjectGroups(groupProjectIds, projectIds, projectsGroups, models, sqlClientPool, project)
+        }
+      }
+    }
+  }
+}
+
+const checkBulkProjectGroupAssociation = async (oid, pid, projectsToMove, groupsToMove, projectsInOtherOrgs, groupsInOtherOrgs, sqlClientPool, models) => {
+  const groupProjectIds = [];
+  const projectIds = [];
+  const projectsGroups = []
+  await checkProjectGroups(groupProjectIds, projectIds, projectsGroups, models, sqlClientPool, pid)
+
+  // for all the projects in the first projects group, iterate through the projects and the groups attached
+  // to these projects and try to build out a map of all the groups and projects that are linked by the primary project
+  if (groupProjectIds.length > 0) {
+    for (const pGroup of groupProjectIds) {
+      const project = await projectHelpers(sqlClientPool).getProjectById(pGroup.project)
+      // check if the project is already in the requested organization
+      if (project.organization != oid && project.organization == null) {
+        let alreadyAdded = false
+        for (const f of projectsToMove) {
+          if (f.id == project.id) {
+            alreadyAdded = true
+          }
+        }
+        if (!alreadyAdded) {
+          // if it isn't already in the requested organization, add it to the list of projects that should be moved
+          projectsToMove.push(project)
+        }
+      } else {
+        // if the project is in a completely different organization
+        if (project.organization != oid) {
+          let alreadyAdded = false
+          for (const f of projectsInOtherOrgs) {
+            if (f.id == project.id) {
+              alreadyAdded = true
+            }
+          }
+          if (!alreadyAdded) {
+            // add it to the lsit of projects that will cause this check to fail
+            projectsInOtherOrgs.push(project)
+          }
+        }
+      }
+    }
+    for (const group of projectsGroups) {
+      // for every group that the project is in, check if the group is already in the requested organization
+      if (group.organization != oid && group.organization == null) {
+        let alreadyAdded = false
+        for (const f of groupsToMove) {
+          if (f.id == group.id) {
+            alreadyAdded = true
+          }
+        }
+        if (!alreadyAdded) {
+          // if it isn't already in the requested organization, add it to the list of groups that should be moved
+          groupsToMove.push(group)
+        }
+      } else {
+        // if the group is in a completely different organization
+        if (group.organization != oid) {
+          let alreadyAdded = false
+          for (const f of groupsInOtherOrgs) {
+            if (f.id == group.id) {
+              alreadyAdded = true
+            }
+          }
+          if (!alreadyAdded) {
+            // add it to the lsit of projects that will cause this check to fail
+            groupsInOtherOrgs.push(group)
+          }
+        }
+      }
+    }
+  }
+}
+
+export const checkBulkImportProjectsAndGroupsToOrganization: ResolverFn = async (
+  _root,
+  { input },
+  { sqlClientPool, models, hasPermission }
+) => {
+  let pid = input.project;
+  let oid = input.organization;
+
+  // platform admin only as it potentially reveals information about projects/orgs/groups
+  await hasPermission('organization', 'add');
+
+  const projectsToMove = []
+  const groupsToMove = []
+  const projectsInOtherOrgs = []
+  const groupsInOtherOrgs = []
+
+  // get all the groups the requested project is in
+  await checkBulkProjectGroupAssociation(oid, pid, projectsToMove, groupsToMove, projectsInOtherOrgs, groupsInOtherOrgs, sqlClientPool, models)
+
+  return { projects: projectsToMove, groups: groupsToMove, otherOrgProjects: projectsInOtherOrgs, otherOrgGroups: groupsInOtherOrgs };
+};
+
+// given a project, collect all the groups that this project has, and all the projects that those groups have and their associated projects
+// and import them into the given organization
+export const bulkImportProjectsAndGroupsToOrganization: ResolverFn = async (
+  root,
+  { input, detachNotifications },
+  { sqlClientPool, hasPermission, userActivityLogger, models }
+) => {
+
+  let pid = input.project;
+  let oid = input.organization;
+
+  // platform admin only as it potentially reveals information about projects/orgs/groups
+  await hasPermission('organization', 'add');
+
+  const projectsToMove = []
+  const groupsToMove = []
+  const projectsInOtherOrgs = []
+  const groupsInOtherOrgs = []
+
+  // get all the groups the requested project is in
+  await checkBulkProjectGroupAssociation(oid, pid,  projectsToMove, groupsToMove, projectsInOtherOrgs, groupsInOtherOrgs, sqlClientPool, models)
+
+  // if anything comes back in projectsInOtherOrgs or groupsInOtherOrgs, then this mutation should fail and inform the user
+  // to run the query first and return the fields that contain information about why it can't move the projects
+  if (projectsInOtherOrgs.length > 0 || groupsInOtherOrgs.length > 0) {
+    throw new Error(
+      `The process detected projects or groups that are in another organization already, you should run checkBulkImportProjectsAndGroupsToOrganization and return otherOrgProjects and otherOrgGroups fields`
+    )
+  }
+
+  // update all projects to be in the organization
+  const groupsDone = [];
+  const projectsDone = [];
+  for (const group of groupsToMove) {
+    // update the groups of the project to be in the organization
+    if (!groupsDone.includes(group.id)) {
+      if (group.organization != oid && group.organization == null) {
+        await models.GroupModel.updateGroup({
+          id: group.id,
+          name: group.name,
+          attributes: {
+            ...group.attributes,
+            "lagoon-organization": [input.organization]
+          }
+        });
+        if (input.organization) {
+          await groupHelpers(sqlClientPool).addOrganizationToGroup(input.organization, group.id)
+        }
+        groupsDone.push(group.id)
+
+        // log this activity
+        userActivityLogger(`User added a group to organization`, {
+          project: '',
+          organization: input.organization,
+          event: 'api:addGroupToOrganization',
+          payload: {
+            data: {
+              group: group.name,
+              organization: oid
+            }
+          }
+        });
+      }
+    }
+  }
+  for (const project of projectsToMove) {
+    if (!projectsDone.includes(project.id)) {
+      if (project.organization != oid && project.organization == null) {
+        if (detachNotifications) {
+          // remove all notifications from projects before adding them to the organizations
+          try {
+            await notificationHelpers(sqlClientPool).removeAllNotificationsFromProject({project: project.id})
+            userActivityLogger(`User removed all notifications from project`, {
+              project: '',
+              organization: input.organization,
+              event: 'api:removeNotificationsFromProject',
+              payload: {
+                data: {
+                  project: project.id,
+                  patch:{
+                    organization: oid,
+                  }
+                }
+              }
+            });
+          } catch (err) {
+            throw new Error(
+              `Unable to remove all notifications from the project`
+            )
+          }
+        }
+
+        // set project.organization
+        await query(
+          sqlClientPool,
+          Sql.updateProjectOrganization({
+            pid: project.id,
+            patch:{
+              organization: oid,
+            }
+          })
+        );
+        projectsDone.push(project.id)
+
+        // log this activity
+        userActivityLogger(`User added a project to organization`, {
+          project: '',
+          organization: input.organization,
+          event: 'api:addExistingProjectToOrganization',
+          payload: {
+            data: {
+              project: project.id,
+              patch:{
+                organization: oid,
+              }
+            }
+          }
+        });
+      }
+    }
+  }
+
+  return { projects: projectsToMove, groups: groupsToMove, otherOrgProjects: projectsInOtherOrgs, otherOrgGroups: groupsInOtherOrgs };
+}
