@@ -125,11 +125,6 @@ export const Sql = {
       routerPattern,
       openshift,
       openshiftProjectPattern,
-      activeSystemsDeploy = "lagoon_controllerBuildDeploy",
-      activeSystemsPromote = "lagoon_controllerBuildDeploy",
-      activeSystemsRemove = "lagoon_controllerRemove",
-      activeSystemsTask = "lagoon_controllerJob",
-      activeSystemsMisc = "lagoon_controllerMisc",
       branches = "true",
       pullrequests = "true",
       productionEnvironment,
@@ -142,8 +137,8 @@ export const Sql = {
       storageCalc = 1,
       problemsUi = 0,
       factsUi = 0,
-      productionBuildPriority = 5,
-      developmentBuildPriority = 6,
+      productionBuildPriority = 6,
+      developmentBuildPriority = 5,
       deploymentsDisabled = 0,
       developmentEnvironmentsLimit = 5,
       organization,
@@ -159,11 +154,6 @@ export const Sql = {
     privateKey,
     subfolder,
     routerPattern,
-    activeSystemsDeploy,
-    activeSystemsPromote,
-    activeSystemsRemove,
-    activeSystemsTask,
-    activeSystemsMisc,
     branches,
     productionEnvironment,
     productionRoutes,
