@@ -49,23 +49,35 @@ export const addSshKey: ResolverFn = async (
 
   const user = await models.UserModel.loadUserByIdOrUsername({
     id: R.prop('id', userInput),
-    username: R.prop('email', userInput)
+    email: R.prop('email', userInput)
   });
 
   await hasPermission('ssh_key', 'add', {
     users: [user.id]
   });
 
-  const { insertId } = await query(
-    sqlClientPool,
-    Sql.insertSshKey({
-      id,
-      name,
-      keyValue,
-      keyType,
-      keyFingerprint: getSshKeyFingerprint(keyFormatted)
-    })
-  );
+  let insertId: number;
+  try {
+     ({insertId} = await query(
+      sqlClientPool,
+      Sql.insertSshKey({
+        id,
+        name,
+        keyValue,
+        keyType,
+        keyFingerprint: getSshKeyFingerprint(keyFormatted)
+      })
+    ));
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error adding SSH key. Key already exists.`
+      );
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   await query(
     sqlClientPool,
     Sql.addSshKeyToUser({ sshKeyId: insertId, userId: user.id })
@@ -130,18 +142,29 @@ export const updateSshKey: ResolverFn = async (
     keyFingerprint = getSshKeyFingerprint(keyFormatted);
   }
 
-  await query(
-    sqlClientPool,
-    Sql.updateSshKey({
-      id,
-      patch: {
-        name,
-        keyType,
-        keyValue,
-        keyFingerprint
-      }
-    })
-  );
+  try {
+    await query(
+      sqlClientPool,
+      Sql.updateSshKey({
+        id,
+        patch: {
+          name,
+          keyType,
+          keyValue,
+          keyFingerprint
+        }
+      })
+    );
+  } catch(error) {
+    if(error.text.includes("Duplicate entry")){
+      throw new Error(
+        `Error updating SSH key. Key already exists.`
+      );
+    } else {
+      throw new Error(error.message);
+    }
+  };
+
   const rows = await query(sqlClientPool, Sql.selectSshKey(id));
 
   userActivityLogger(`User updated ssh key '${id}'`, {
