@@ -1,18 +1,16 @@
-// @ts-ignore
 import * as R from 'ramda';
 import { verify } from 'jsonwebtoken';
 import { logger } from '../loggers/logger';
 import { getConfigFromEnv } from '../util/config';
 import { isNotNil } from './func';
 import { keycloakGrantManager } from '../clients/keycloakClient';
-// @ts-ignore
 const { userActivityLogger } = require('../loggers/userActivityLogger');
 import { Group } from '../models/group';
 import { User } from '../models/user';
-import { saveRedisKeycloakCache } from '../clients/redisClient';
 
 interface ILegacyToken {
   iat: string;
+  exp: string;
   iss: string;
   sub: string;
   aud: string;
@@ -105,7 +103,26 @@ export const getCredentialsForLegacyToken = async token => {
     throw new Error('Decoding token resulted in "null" or "undefined".');
   }
 
-  const { role = 'none', aud, sub, iss, iat } = decoded;
+  const { role = 'none', aud, sub, iss, iat, exp } = decoded;
+
+  // check the expiration on legacy tokens, reject them if necessary
+  const maxExpiry = getConfigFromEnv('LEGACY_EXPIRY_MAX', '3600') // 1hour default
+  const rejectLegacyExpiry = getConfigFromEnv('LEGACY_EXPIRY_REJECT', 'false') // don't reject intially, just log
+  if (exp) {
+    if ((parseInt(exp)-parseInt(iat)) > parseInt(maxExpiry)) {
+      const msg = `Legacy token (sub:${sub}; iss:${iss}) expiry ${(parseInt(exp)-parseInt(iat))} is greater than ${parseInt(maxExpiry)}`
+      logger.warn(msg);
+      if (rejectLegacyExpiry == "true") {
+        throw new Error(msg);
+      }
+    }
+  } else {
+    const msg = `Legacy token (sub:${sub}; iss:${iss}) has no expiry`
+    logger.warn(msg);
+    if (rejectLegacyExpiry == "true") {
+      throw new Error(msg);
+    }
+  }
 
   if (aud !== getConfigFromEnv('JWTAUDIENCE')) {
     throw new Error('Token audience mismatch.');
@@ -164,6 +181,7 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
       userGroupRole?: [string];
       organizationQuery?: [string];
       userOrganizations?: [string];
+      userOrganizationsAdmin?: [string];
       userOrganizationsView?: [string];
     } = {
       currentUser: [currentUser.id]
@@ -189,6 +207,13 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
       claims = {
         ...claims,
         userOrganizations: [`${R.prop('lagoon-organizations', currentUser.attributes)}`]
+      };
+    }
+    // check organization admin attributes
+    if (R.prop('lagoon-organizations-admin', currentUser.attributes)) {
+      claims = {
+        ...claims,
+        userOrganizationsAdmin: [`${R.prop('lagoon-organizations-admin', currentUser.attributes)}`]
       };
     }
     // check organization viewer attributes
@@ -224,7 +249,7 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
           // but could happen elsewhere
           const keycloakUsersGroups = await UserModel.getAllGroupsForUser(currentUser.id);
           // grab the users project ids and roles in the first request
-          groupRoleProjectIds = await UserModel.getAllProjectsIdsForUser(currentUser, keycloakUsersGroups);
+          groupRoleProjectIds = await UserModel.getAllProjectsIdsForUser(currentUser.id, keycloakUsersGroups);
 
           [highestRoleForProject, upids] = getUserRoleForProjectFromRoleProjectIds(groupRoleProjectIds, projectId)
         }
@@ -306,7 +331,6 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
       authzRequest = {
         ...authzRequest,
         claim_token_format: 'urn:ietf:params:oauth:token-type:jwt',
-        // @ts-ignore
         claim_token: Buffer.from(JSON.stringify(claims)).toString('base64')
       };
     }
@@ -319,7 +343,6 @@ export const keycloakHasPermission = (grant, requestCache, modelClients, service
     };
 
     try {
-      // @ts-ignore
       const newGrant = await keycloakGrantManager.checkPermissions(
         authzRequest,
         request
