@@ -8,4 +8,4 @@ iat = datetime.now(timezone.utc)
 exp = iat + timedelta(minutes=1)
 payload = {'exp': exp, 'iat': iat, 'role': 'admin', 'aud': os.environ['JWTAUDIENCE'], 'sub': 'ssh'}
 
-print(jwt.encode(payload, os.environ['JWTSECRET'], algorithm='HS256').decode())
+print(jwt.encode(payload, os.environ['JWTSECRET'], algorithm='HS256'))
