@@ -47,7 +47,12 @@ export async function bitbucketPullRequestUpdated(webhook: WebhookRequestData, p
     }
 
     let buildName = generateBuildId();
-
+    // try get the user from the webhook payload
+    // otherwise just use "webhook" as the trigger user
+    let sourceUser = "webhook"
+    if (body.pullrequest.author.username) {
+      sourceUser = body.pullrequest.author.username
+    }
     const data: deployData = {
       repoName: body.repository.full_name,
       repoUrl: body.repository.links.html.href,
@@ -61,7 +66,11 @@ export async function bitbucketPullRequestUpdated(webhook: WebhookRequestData, p
       baseBranchName: baseBranchName,
       baseSha: baseSha,
       branchName: `pr-${body.pullrequest.id}`,
-      buildName: buildName
+      buildName: buildName,
+      sourceUser: sourceUser,
+      sourceType: "WEBHOOK",
+      bulkId: webhook.bulkId,
+      bulkName: webhook.bulkName,
     }
 
     try {
@@ -73,8 +82,6 @@ export async function bitbucketPullRequestUpdated(webhook: WebhookRequestData, p
     } catch (error) {
       switch (error.name) {
         case "ProjectNotFound":
-        case "NoActiveSystemsDefined":
-        case "UnknownActiveSystem":
           // These are not real errors and also they will happen many times. We just log them locally but not throw an error
           sendToLagoonLogs('info', project.name, uuid, `${webhooktype}:${event}:handledButNoTask`, meta,
             `*[${project.name}]* PR ${body.object_attributes.id} updated. No deploy task created, reason: ${error}`
