@@ -1,7 +1,7 @@
 import { ChannelWrapper } from 'amqp-connection-manager';
 import { ConsumeMessage } from 'amqplib';
 import { logger } from '@lagoon/commons/dist/logs/local-logger';
-import { getProjectsByGitUrl } from '@lagoon/commons/dist/api';
+import { Project as LagoonProject, ProjectNotFound, lagoonApi } from '@lagoon/commons/dist/api';
 import { sendToLagoonLogs } from '@lagoon/commons/dist/logs/lagoon-logger';
 import { githubPullRequestClosed } from '../handlers/githubPullRequestClosed';
 import { githubPullRequestOpened } from '../handlers/githubPullRequestOpened';
@@ -22,6 +22,7 @@ import { gitlabBranchDeleted } from '../handlers/gitlabBranchDeleted';
 import { gitlabPullRequestClosed } from '../handlers/gitlabPullRequestClosed';
 import { gitlabPullRequestOpened } from '../handlers/gitlabPullRequestOpened';
 import { gitlabPullRequestUpdated } from '../handlers/gitlabPullRequestUpdated';
+import crypto from 'crypto';
 
 import {
   WebhookRequestData,
@@ -33,13 +34,31 @@ export async function processProjects(
   channelWrapperWebhooks: ChannelWrapper
 ): Promise<void> {
   const webhook: WebhookRequestData = JSON.parse(rabbitMsg.content.toString());
-
-  let projects: Project[];
-
   const { webhooktype, event, giturl, uuid, body } = webhook;
 
   try {
-    projects = await getProjectsByGitUrl(giturl);
+    const result = await lagoonApi.query(
+      `
+      query webhookProcessProjects($giturl: String!) {
+        allProjects(gitUrl: $giturl) {
+          name
+          deploymentsDisabled
+        }
+      }
+    `,
+      { giturl }
+    ) as {
+      allProjects: Pick<
+        LagoonProject,
+        'name' | 'deploymentsDisabled'
+      >[];
+    }
+
+    if (!result || !result.allProjects || !result.allProjects.length) {
+      throw new ProjectNotFound(`Cannot find project for git repo ${giturl}`);
+    }
+
+    var projects = result.allProjects;
   } catch (error) {
     if (error.name == 'ProjectNotFound') {
       const meta = {
@@ -120,6 +139,15 @@ export async function processProjects(
       channelWrapperWebhooks.ack(rabbitMsg);
     }
     return;
+  }
+
+  // if there are more than 1 projects returned, it is probably a polysite deployment
+  if (projects.length > 1) {
+    // label them as a bulk or grouped deployment
+    // assign a random uuid
+    webhook.bulkId = crypto.randomUUID()
+    // and then add the name as polysite indicating what type of event started it
+    webhook.bulkName = `Polysite - ${webhooktype}:${event} - ${new Date().toISOString()}`
   }
 
   projects.forEach(async project => {
